refactor: move hooks to hooks file
